refactor(amazonq): move local project context controller to shared
