+++ server/aws-lsp-codewhisperer/src/shared/localProjectContextController.test.ts
